Refactored ViewModels into their own folders
