+++ src/GeekPlatform/Controllers/JelentkezesController.cs
@@ -8,7 +8,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.AspNetCore.Identity;
 using GeekPlatform.Models;
-using GeekPlatform.Models.ViewModels;
+using GeekPlatform.ViewModels.Jelentkezes;
 
 // For more information on enabling MVC for empty projects, visit http://go.microsoft.com/fwlink/?LinkID=397860
