+++ src/GeekPlatform/ViewModels/Jelentkezes/JelekezesViewModel.cs
@@ -4,7 +4,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace GeekPlatform.Models.ViewModels
+using GeekPlatform.Models;
+
+namespace GeekPlatform.ViewModels.Jelentkezes
 {
     public class JelekezesViewModel : ViewModelBase
     {
