+++ src/GeekPlatform/ViewModels/ViewModelBase.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace GeekPlatform.Models.ViewModels
+namespace GeekPlatform.ViewModels
 {
     public class ViewModelBase
     {
