+++ src/GeekPlatform/Views/Jelentkezes/Index.cshtml
@@ -1,7 +1,7 @@
 ﻿@*
     For more information on enabling MVC for empty projects, visit http://go.microsoft.com/fwlink/?LinkID=397860
 *@
-@using GeekPlatform.Models.ViewModels;
+@using GeekPlatform.ViewModels.Jelentkezes;
